Refactored graph package to tree package.
