+++ PL2/PL2-gui/src/main/java/nl/tudelft/pl2016gr2/gui/view/tree/TreeController.java
@@ -1,4 +1,4 @@
-package nl.tudelft.pl2016gr2.gui.view.graph;
+package nl.tudelft.pl2016gr2.gui.view.tree;
 
 import javafx.animation.Timeline;
 import javafx.scene.control.Button;
